unify some code
