+++ roc_picker/delta_functions.py
@@ -1,6 +1,7 @@
 import functools, numpy as np, scipy.optimize
+from .discrete_base import DiscreteROCBase
 
-class DeltaFunctions:
+class DeltaFunctions(DiscreteROCBase):
   def __init__(self, responders, nonresponders, *, flip_sign=False):
     self.responders = responders
     self.nonresponders = nonresponders
